backend/requirements.txt
Flask==3.1.0
Werkzeug==3.1.3
PyJWT==2.10.1
cryptography==44.0.0
